requirements.txt
flask
flask-socketio
python-socketio
python-engineio
pywebview
pyinstaller
pywin32; platform_system == "Windows"
